LPS-64471 unify upgrade processes again to avoid increasing package version
