+++ portal-impl/src/com/liferay/portal/upgrade/v7_0_0/UpgradeAsset.java
@@ -14,15 +14,23 @@
 
 package com.liferay.portal.upgrade.v7_0_0;
 
+import com.liferay.asset.kernel.model.AssetCategoryConstants;
 import com.liferay.portal.kernel.dao.jdbc.AutoBatchPreparedStatementUtil;
 import com.liferay.portal.kernel.upgrade.UpgradeProcess;
+import com.liferay.portal.kernel.util.ArrayUtil;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.LoggingTimer;
 import com.liferay.portal.kernel.util.PortalUtil;
 import com.liferay.portal.kernel.util.StringBundler;
+import com.liferay.portal.kernel.util.StringUtil;
+import com.liferay.portal.kernel.util.UnicodeProperties;
 import com.liferay.portal.upgrade.v7_0_0.util.AssetEntryTable;
+import com.liferay.portlet.asset.util.AssetVocabularySettingsHelper;
 
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
+
+import java.util.Arrays;
 
 /**
  * @author Gergely Mathe
@@ -37,6 +45,7 @@ public class UpgradeAsset extends UpgradeProcess {
 			new AlterColumnType("summary", "TEXT null"));
 
 		updateAssetEntries();
+		updateAssetVocabularies();
 	}
 
 	protected long getDDMStructureId(String structureKey) throws Exception {
@@ -99,6 +108,65 @@ public class UpgradeAsset extends UpgradeProcess {
 				}
 			}
 		}
+	}
+
+	protected void updateAssetVocabularies() throws Exception {
+		try (LoggingTimer loggingTimer = new LoggingTimer();
+			PreparedStatement ps1 = connection.prepareStatement(
+				"select vocabularyId, settings_ from AssetVocabulary");
+			PreparedStatement ps2 =
+				AutoBatchPreparedStatementUtil.concurrentAutoBatch(
+					connection,
+					"update AssetVocabulary set settings_ = ? where " +
+						"vocabularyId = ?");
+			ResultSet rs = ps1.executeQuery()) {
+
+			while (rs.next()) {
+				long vocabularyId = rs.getLong("vocabularyId");
+				String settings = rs.getString("settings_");
+
+				ps2.setString(1, upgradeVocabularySettings(settings));
+				ps2.setLong(2, vocabularyId);
+
+				ps2.addBatch();
+			}
+
+			ps2.executeBatch();
+		}
+	}
+
+	protected String upgradeVocabularySettings(String settings) {
+		UnicodeProperties properties = new UnicodeProperties(true);
+
+		properties.fastLoad(settings);
+
+		AssetVocabularySettingsHelper vocabularySettingsHelper =
+			new AssetVocabularySettingsHelper();
+
+		vocabularySettingsHelper.setMultiValued(
+			GetterUtil.getBoolean(properties.getProperty("multiValued"), true));
+
+		long[] classNameIds = StringUtil.split(
+			properties.getProperty("selectedClassNameIds"), 0L);
+
+		long[] classTypePKs = new long[classNameIds.length];
+
+		Arrays.fill(classTypePKs, AssetCategoryConstants.ALL_CLASS_TYPE_PK);
+
+		long[] requiredClassNameIds = StringUtil.split(
+			properties.getProperty("requiredClassNameIds"), 0L);
+
+		boolean[] requireds = new boolean[classNameIds.length];
+
+		for (int i = 0; i < classNameIds.length; i++) {
+			requireds[i] = ArrayUtil.contains(
+				requiredClassNameIds, classNameIds[i]);
+		}
+
+		vocabularySettingsHelper.setClassNameIdsAndClassTypePKs(
+			classNameIds, classTypePKs, requireds);
+
+		return vocabularySettingsHelper.toString();
 	}
 
 }
